+++ src/components/app/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Route } from 'react-router-dom'
 import { Layout } from 'antd'
-import Board from '../board/Board'
+import DndBoard from '../board/DndBoard'
 import './style.css'
 
 const { Content } = Layout
@@ -9,7 +9,7 @@ const { Content } = Layout
 const App = () => (
   <Layout className="appLayout">
     <Content className="appContent">
-      <Route exact path="/" component={Board} />
+      <Route exact path="/" component={DndBoard} />
     </Content>
   </Layout>
 )
